implement commandTime-based smoothing.
the basic idea is that server sends in playerState the command time, whose delta can vary causing jumpiness.
although server executes X fps, client sends position updates at Y fps and server roughly respects that, hence the jumpiness.
this patch uses the commandTime rather than snapshot time to lerp player position changes, resulting in much smoother playback.

included in this patch:
initial version of commandTime-based smoothing
v2 of commandTime-based interpolation.  store playerStates in separate struct, and create struct to start doing the same for other players if possible.
add support for one frame lookahead
add cvar to allow disabling of commandTime-based smoothing
fixes connecting to remote server
fix for teleports when using command smoothing
fixes jumpy speed trail when using command smoothing

fix corpse holding duplicate weapon that it just threw.

issue is that due to command smoothing, eflags can be a frame behind. weapon clearing happens in an event
and eflags are checked to prevent it from immediately respawning. but in command smoothing, it can be cleared
before the ef_dead flag is set, causing it to be immediately reattached. the fix is to use the health from
snap as if we were not smoothing so that ef_dead is set as it would normally be.